+++ index.js
@@ -1,32 +1,47 @@
 'use strict';
+
 var _ = require('lodash');
 var jsonServer = require('json-server');
+var Merger = require('./merger');
+var through = require('through2');
 var utils = require('gulp-util');
-var fs = require('fs');
+var chalk = require('chalk');
 var bodyParser = require('body-parser');
+var enableDestroy = require('server-destroy');
 
 var GulpJsonServer = function(options){
 	this.server = null;
 	this.instance = null;
 	this.router = null;
 	this.serverStarted = false;
+	this.devMode = false;
 
 	this.options = {
-		data: 'db.json',
 		port: 3000,
 		rewriteRules: null,
 		customRoutes: null,
 		baseUrl: null,
 		id: 'id',
-		deferredStart: false,
-		static: null
+		static: null,
+		cumulative: false,
+		cumulativeSession: true,
+		debug: false
 	};
-	_.assign(this.options, options || {});
+	
+	var self = this;
 
-	this.start = function () {
+	_.extend(this.options, options || {});
+
+	var start = function (data) {
 		if(this.serverStarted){
-			utils.log('JSON server already started');
+			if(this.options.debug){
+				console.log(chalk.yellow('JSON server already started'));
+			}
 			return this.instance;
+		}
+
+		if(this.options.debug){
+			console.log(chalk.green("starting server"));
 		}
 
         var server = jsonServer.create();
@@ -51,7 +66,7 @@ var GulpJsonServer = function(options){
 			}
 		}
 
-		var router = jsonServer.router(this.options.data);
+		var router = jsonServer.router(data || this.options.data);
 		if(this.options.baseUrl) {
 			server.use(this.options.baseUrl, router);
 		}
@@ -60,73 +75,120 @@ var GulpJsonServer = function(options){
 		}
 
 		if(this.options.id){
-			router.db._.id = this.options.id;
+			var newId = this.options.id;
+			router.db._.mixin({
+				__id: function(){
+					return newId;
+				}
+			});
 		}
 
 		this.server = server;
 		this.router = router;
 		this.instance = server.listen(this.options.port);
+
+		enableDestroy(this.instance);
+
 		this.serverStarted = true;
 
 		return this.instance;
-	};
-
-	var ensureServerStarted = function(){
-		if(this.instance === null){
-			throw 'JSON server not started';
-		}
 	}.bind(this);
 
-	this.kill = function(){
-		ensureServerStarted();
-		this.instance.close();
+	var reload = function(data){
+		if(typeof data === 'undefined'){
+			if(this.options.debug){
+				console.log(chalk.yellow('nothing to reload, quit'));
+			}
+			return;
+		}
+
+		if(this.options.debug){
+			console.log(chalk.green("reloading data:"));			
+			console.log(JSON.stringify(data));
+			console.log(chalk.yellow("destroying server..."));
+		}
+		var gulpJsonSrvInstance = this;
+		this.kill(function(){
+			if(gulpJsonSrvInstance.options.debug){
+				console.log(chalk.yellow("server destroyed"));
+			}
+		});
+		start(data);
+	}.bind(this);
+
+
+
+	this.kill = function(callback){
+		if(this.instance){
+			this.instance.destroy(callback);
+			this.serverStarted = false;
+		}
 	};
+	
+	this.pipe = function(options){
+		var isCumulative = options && typeof(options.cumulative) !== "undefined" ? options.cumulative : self.options.cumulative;
+		var isCumulativeSession = options && typeof(options.cumulativeSession) !== "undefined" ? options.cumulativeSession : self.options.cumulativeSession;
 
-	this.reload = function(data){
-		ensureServerStarted();
+		// HACK json-server to get its db object if needed
+		var aggregatorObject = self.serverStarted && isCumulative ? self.router.db.getState() || {} : {};
+		
+		if(this.devMode){
+			console.log(chalk.red("server started: " + this.serverStarted));			
+			console.log(chalk.red("cumulative: " + this.options.cumulative));			
+			console.log(chalk.red("aggregator object:"));			
+			console.log(JSON.stringify(aggregatorObject));
+		}
 
-		var isDataFile = typeof this.options.data === 'string';
-		var newDb = null;
-
-		if(typeof(data) === 'undefined'){
-			if(!isDataFile){
-				// serving in-memory DB, exit without changes
+		return through.obj(function (file, enc, cb) {
+			if (file.isNull()) {
+				cb(null, file);
 				return;
 			}
 
-			utils.log('reload from default file', utils.colors.yellow(this.options.data));
-			newDb = JSON.parse(fs.readFileSync(this.options.data));
-		}
-		else if(data !== null){
-			if(typeof data === 'string'){
-				// attempt to reload file
-				utils.log('reload from file', utils.colors.yellow(data));
-				newDb = JSON.parse(fs.readFileSync(data));
+			if (file.isStream()) {
+				cb(new utils.PluginError('gulp-json-srv', 'Streaming not supported'));
+				return;
 			}
-			else{
-				// passed new DB object, store it
-				utils.log('reload from object');
-				newDb = data;
+
+			try {
+				var appendedObject = JSON.parse(file.contents.toString());
+				if(self.options.debug){
+					console.log(chalk.green('file data:'));
+					console.log(JSON.stringify(appendedObject));
+				}
+
+				if(isCumulativeSession){
+					aggregatorObject = new Merger(self.options).merge(aggregatorObject, appendedObject || {});
+					if(self.options.debug){
+						console.log(chalk.green("combine DB data in session"));
+					}
+				}
+				else{
+					aggregatorObject = appendedObject || {};
+					if(self.options.debug){
+						console.log(chalk.green("override DB data in session (cumulativeSession=false)"));
+					}
+				}
+				
+				if(self.devMode){
+					console.log(chalk.red("pipe reloading data:"));			
+					console.log(JSON.stringify(aggregatorObject));
+				}
+				
+				reload(aggregatorObject);
+
+				this.push(file);
+			} catch (err) {
+				this.emit('error', new utils.PluginError('gulp-json-srv', err));
 			}
-		}
 
-		if(newDb === null){
-			throw 'No valid data passed for reloading. You should pass either data file path or new DB in-memory object';
-		}
-
-		this.router.db.object = newDb;
-		utils.log(utils.colors.magenta('server reloaded'));
+			cb();
+		});
 	};
-
-	// ==== Initialization ====
-	if(!this.options.deferredStart){
-		this.start();
-	}
 };
 
-
 module.exports = {
-	start: function(options){
+	create: function(options){
 		return new GulpJsonServer(options);
-	}
+	}	
 };