+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Release notes
 
+### v1.0.0 - breaking changes
+* Whole plugin redesigned to meet the requirements of gulp plugin guidelines.
+* Added ability to pipe files to the plugin.
+* 'data' and 'deferredStart' removed.
+* 'debug', 'cumulative' and 'cumulativeSession' options added.
+* Plugin do not take objects as input anymore, start() and reload() methods removed, the only way to pass data to plugin is via pipeline.
+
 ### v0.2.0
 * Added ability to add rewrite rules to static files.
 * Added ability to use custom routes (reflect json-server changes)
@@ -33,4 +40,4 @@ Basic version of plugin with ability to start json-server from specified file or
 
 ## License
 
-MIT © Nikita Ivanov (http://borodatik.net)
+MIT © 2016 Nikita Ivanov